requirements.txt
absl-py==0.13.0
alabaster==0.7.12
appdirs==1.4.3
apturl==0.5.2
astunparse==1.6.3
awscli==1.18.69
Babel==2.9.1
backcall==0.1.0
bcrypt==3.1.7
blinker==1.4
boto3==1.18.10
botocore==1.21.10
Brlapi==0.7.0
cachetools==4.2.2
certifi==2019.11.28
chardet==3.0.4
click==8.0.1
colorama==0.4.3
command-not-found==0.3
coverage==5.5
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
esb-common==0.1
fasteners==0.14.1
filelock==3.0.12
Flask==2.0.1
future==0.18.2
gast==0.3.3
gitdb==4.0.7
GitPython==0.3.6
google-auth==1.31.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.38.0
h5py==2.10.0
httplib2==0.14.0
idna==2.8
imagesize==1.2.0
importlib-metadata==1.5.0
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.15.2
Jinja2==3.0.1
jmespath==0.9.4
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.6.3
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
monotonic==1.5
more-itertools==4.2.0
mxnet==1.6.0
netifaces==0.10.4
networkx==2.5.1
nose==1.3.7
numpy==1.18.5
oauthlib==3.1.0
olefile==0.46
onnx==1.9.0
opt-einsum==3.3.0
packaging==21.0
paramiko==2.6.0
parso==0.5.2
pdftotext==2.1.6
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
prettytable==2.1.0
prompt-toolkit==2.0.10
protobuf==3.17.3
psutil==5.8.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.5.0
pycups==1.9.73
pydocstyle==2.1.1
Pygments==2.9.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pypi-publisher==0.0.4
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-docx==0.8.11
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
roman==2.0.0
rsa==4.7.2
s3transfer==0.5.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
smmap==4.0.0
snowballstemmer==2.0.0
Sphinx==4.0.3
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
systemd-python==234
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.3
tensorflow-estimator==2.3.0
termcolor==1.1.0
test-generator==0.1.1
traitlets==4.3.3
typing-extensions==3.10.0.0
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.1.8
Werkzeug==2.0.1
wget==3.2
wrapt==1.12.1
xkit==0.0.0
zipp==1.0.0
